# Make session serializer extensible