--- a/tests/test_basic.py
+++ b/tests/test_basic.py
@@ -435,34 +435,31 @@ def test_session_special_types(app, client):
     now = datetime.utcnow().replace(microsecond=0)
     the_uuid = uuid.uuid4()
 
-    @app.after_request
-    def modify_session(response):
-        flask.session['m'] = flask.Markup('Hello!')
-        flask.session['u'] = the_uuid
-        flask.session['dt'] = now
-        flask.session['b'] = b'\xff'
-        flask.session['t'] = (1, 2, 3)
-        flask.session['spacefirst'] = {' t': 'not-a-tuple'}
-        flask.session['withunderscores'] = {' t__': 'not-a-tuple'}
-        flask.session['notadict'] = {' di': 'not-a-dict'}
-        return response
-
     @app.route('/')
     def dump_session_contents():
-        return pickle.dumps(dict(flask.session))
+        flask.session['t'] = (1, 2, 3)
+        flask.session['b'] = b'\xff'
+        flask.session['m'] = flask.Markup('<html>')
+        flask.session['u'] = the_uuid
+        flask.session['d'] = now
+        flask.session['t_tag'] = {' t': 'not-a-tuple'}
+        flask.session['di_t_tag'] = {' t__': 'not-a-tuple'}
+        flask.session['di_tag'] = {' di': 'not-a-dict'}
+        return '', 204
 
-    client.get('/')
-    rv = pickle.loads(client.get('/').data)
-    assert rv['m'] == flask.Markup('Hello!')
-    assert type(rv['m']) == flask.Markup
-    assert rv['dt'] == now
-    assert rv['u'] == the_uuid
-    assert rv['b'] == b'\xff'
-    assert type(rv['b']) == bytes
-    assert rv['t'] == (1, 2, 3)
-    assert rv['spacefirst'] == {' t': 'not-a-tuple'}
-    assert rv['withunderscores'] == {' t__': 'not-a-tuple'}
-    assert rv['notadict'] == {' di': 'not-a-dict'}
+    with client:
+        client.get('/')
+        s = flask.session
+        assert s['t'] == (1, 2, 3)
+        assert type(s['b']) == bytes
+        assert s['b'] == b'\xff'
+        assert type(s['m']) == flask.Markup
+        assert s['m'] == flask.Markup('<html>')
+        assert s['u'] == the_uuid
+        assert s['d'] == now
+        assert s['t_tag'] == {' t': 'not-a-tuple'}
+        assert s['di_t_tag'] == {' t__': 'not-a-tuple'}
+        assert s['di_tag'] == {' di': 'not-a-dict'}
 
 
 def test_session_cookie_setting(app):