@davidism 2017-06-02
@ThiefMaster from https://github.com/pallets/flask/pull/1452#issuecomment-305609953:

Re: performance: new method is ~ 30 µs slower.

~~~
old: 70.7 µs ± 164 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
new: 96.3 µs ± 678 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
~~~

You'd need to handle 2,000,000 requests before you wasted 1 extra minute. There's always Flask-Session, etc. where server-side sessions are serialized with `pickle`, which is *much* faster.